--- models/Comment.js
@@ -1,61 +1,0 @@
-const { Model, DataTypes } = require('sequelize');
-const sequelize = require('../config/connection');
-const User = require('./User'); // Import the User model
-const Medication = require('./Medication'); // Import the Medication model
-
-class Comment extends Model {}
-
-Comment.init(
-  {
-    id: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      primaryKey: true,
-      autoIncrement: true,
-    },
-    content: {
-      type: DataTypes.TEXT,
-      allowNull: false,
-      validate: {
-        len: [1, 500] // Example validation: content length between 1 and 500 characters
-      }
-    },
-    createdAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-    },
-    updatedAt: {
-      type: DataTypes.DATE,
-      allowNull: false,
-    },
-    userId: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      references: {
-        model: User,
-        key: 'id',
-      },
-    },
-    medicationId: {
-      type: DataTypes.INTEGER,
-      allowNull: false,
-      references: {
-        model: Medication,
-        key: 'id',
-      },
-    },
-  },
-  {
-    sequelize,
-    timestamps: true,
-    freezeTableName: true,
-    underscored: true,
-    modelName: 'comment',
-  }
-);
-
-// Define associations
-Comment.belongsTo(User, { foreignKey: 'userId', onDelete: 'CASCADE' });
-Comment.belongsTo(Medication, { foreignKey: 'medicationId', onDelete: 'CASCADE' });
-
-module.exports = Comment;
