+++ public/js/login.js
@@ -12,7 +12,7 @@ const loginFormHandler = async (event) => {
     });
 
     if (response.ok) {
-      document.location.replace('/home'); // GETS the /home ednpoint
+      document.location.replace('/dashboard'); // GETS the /dashboard endpoint
     } else {
       alert('Failed to log in');
     }
@@ -22,3 +22,16 @@ const loginFormHandler = async (event) => {
 document
   .querySelector('.login-form')
   .addEventListener('submit', loginFormHandler);
+
+document.addEventListener("DOMContentLoaded", function () {
+  // Select the Create Account button
+  var createAccountButton = document.getElementById("createAccountButton");
+
+  // Add event listener to the Create Account button
+  createAccountButton.addEventListener("click", function () {
+    // Execute the createAccount.handlebars file or perform any other action you need
+    // For example, you can redirect to a new page using window.location.href
+    window.location.href = "/api/users/createAccount";
+  });
+});
+
