+++ seeds/medicationData.json
@@ -1,67 +1,47 @@
 [
   {
-    "name": "Ampicillin",
-    "dosage": "250mg",
-    "monday": true,
-    "tuesday": true,
-    "wednesday": true,
-    "thursday": true,
-    "friday": false,
-    "saturday": true,
-    "sunday": false,
-    "inUse": true,
-    "userId": "2"
+      "medicationName": "Ampicillin",
+      "strength": "250mg",
+      "dosage": "1 tablet",
+      "frequency": "daily",
+      "route": "Oral",
+      "duration": "indefinite",
+      "user_id": "2"
   },
   {
-    "name": "Ibuprofen",
-    "dosage": "200mg",
-    "monday": true,
-    "tuesday": false,
-    "wednesday": true,
-    "thursday": false,
-    "friday": true,
-    "saturday": false,
-    "sunday": false,
-    "inUse": false,
-    "userId": "3"
+      "medicationName": "Metroprolol",
+      "strength": "250mg",
+      "dosage": "1 tablet",
+      "frequency": "daily",
+      "route": "Oral",
+      "duration": "indefinite",
+      "user_id": "2"
   },
   {
-    "name": "Ciprofloxacin",
-    "dosage": "500mg",
-    "monday": true,
-    "tuesday": true,
-    "wednesday": true,
-    "thursday": true,
-    "friday": true,
-    "saturday": false,
-    "sunday": false,
-    "inUse": true,
-    "userId": "4"
+      "medicationName": "Aspirin",
+      "strength": "81",
+      "dosage": "1 tablet",
+      "frequency": "daily",
+      "route": "Oral",
+      "duration": "indefinite",
+      "user_id": "2"
   },
   {
-    "name": "Aspirin",
-    "dosage": "75mg",
-    "monday": true,
-    "tuesday": true,
-    "wednesday": false,
-    "thursday": true,
-    "friday": false,
-    "saturday": false,
-    "sunday": true,
-    "inUse": false,
-    "userId": "5"
+      "medicationName": "Neosporin",
+      "strength": "250mg",
+      "dosage": "1 tablet",
+      "frequency": "daily",
+      "route": "Skin",
+      "duration": "indefinite",
+      "user_id": "2"
   },
   {
-    "name": "Levothyroxine",
-    "dosage": "50mcg",
-    "monday": true,
-    "tuesday": false,
-    "wednesday": true,
-    "thursday": false,
-    "friday": true,
-    "saturday": false,
-    "sunday": false,
-    "inUse": true,
-    "userId": "1"
+      "medicationName": "Oxycontin",
+      "strength": "250mg",
+      "dosage": "1 tablet",
+      "frequency": "daily",
+      "route": "Oral",
+      "duration": "indefinite",
+      "user_id": "1"
   }
 ]
