--- seeds/patientData.json
@@ -1,87 +1,0 @@
-[
-  {
-    "first_name": "John",
-    "last_name": "Doe",
-    "date_of_birth": "1990-01-01",
-    "gender": "Male",
-    "height": 180.5,
-    "weight": 75,
-    "email": "john1@example.com",
-    "street": "123 Main Street",
-    "apt_unit_number": "Apt 101",
-    "city": "Anytown",
-    "state": "NY",
-    "zip_code": "12345",
-    "phone_1": "555-123-4567",
-    "phone_2": "555-987-6543",
-    "userId": 1
-  },
-  {
-    "first_name": "Jane",
-    "last_name": "Doe",
-    "date_of_birth": "1992-03-15",
-    "gender": "Female",
-    "height": 165.0,
-    "weight": 60,
-    "email": "jane2@example.com",
-    "street": "456 Elm Street",
-    "apt_unit_number": "Unit B",
-    "city": "Sometown",
-    "state": "CA",
-    "zip_code": "54321",
-    "phone_1": "555-987-6543",
-    "phone_2": null,
-    "userId": null
-  },
-  {
-    "first_name": "Michael",
-    "last_name": "Smith",
-    "date_of_birth": "1985-07-10",
-    "gender": "Male",
-    "height": 175.2,
-    "weight": 80,
-    "email": "michael3@example.com",
-    "street": "789 Oak Avenue",
-    "apt_unit_number": null,
-    "city": "Othertown",
-    "state": "TX",
-    "zip_code": "67890",
-    "phone_1": "555-234-5678",
-    "phone_2": null,
-    "userId": null
-  },
-  {
-    "first_name": "Emily",
-    "last_name": "Johnson",
-    "date_of_birth": "1988-09-20",
-    "gender": "Female",
-    "height": 160.0,
-    "weight": 55,
-    "email": "emily4@example.com",
-    "street": "101 Pine Street",
-    "apt_unit_number": "Suite 200",
-    "city": "Cityville",
-    "state": "FL",
-    "zip_code": "45678",
-    "phone_1": "555-345-6789",
-    "phone_2": "555-456-7890",
-    "userId": null
-  },
-  {
-    "first_name": "William",
-    "last_name": "Brown",
-    "date_of_birth": "1995-05-05",
-    "gender": "Male",
-    "height": 185.0,
-    "weight": 90,
-    "email": "william5@example.com",
-    "street": "222 Maple Avenue",
-    "apt_unit_number": null,
-    "city": "Anothertown",
-    "state": "IL",
-    "zip_code": "98765",
-    "phone_1": "555-567-8901",
-    "phone_2": null,
-    "userId": null
-  }
-]
