--- views/addpatient.handlebars
@@ -1,13 +1,0 @@
- <main id="page-container" class="main-container">
-        <!-- New Patient Form Page -->
-        <section class="new-patient-form">
-            <form id="patientForm">
-                <!-- The form fields will be injected here -->
-            </form>
-            <div id="validation-messages"></div>
-            <!-- Form Submission Button -->
-            <button type="submit" class="button submit-button">Submit</button>
-            </form>
-        </section>
-    </main>
-    <script src="./assets/js/new-patient.js"></script> <!-- Adjusted script name for general use -->
