+++ views/createAccount.handlebars
@@ -1,0 +1,13 @@
+<main id="page-container" class="main-container">
+    <!-- New Patient Form Page -->
+    <section class="new-patient-form">
+        <form id="patientForm">
+            <!-- The form fields will be injected here -->
+            <div id="validation-messages"></div>
+        </form>
+        <!-- Form Submission Button -->
+        <button type="submit" class="button submit-button">Submit</button>
+    </section>
+</main>
+<script src="/js/createAccountForm.js"></script>
+<script src="/js/createAccount.js"></script>
