--- views/homepage.handlebars
@@ -1,8 +1,0 @@
-<div id="dayContainer" class="container">
-  <div class="dayBox">S</div>
-  <div class="dayBox">M</div>
-  <div class="dayBox">W</div>
-  <div class="dayBox">T</div>
-  <div class="dayBox">F</div>
-  <div class="dayBox">S</div>
-</div>
