+++ views/layouts/main.handlebars
@@ -12,13 +12,9 @@
 <body>
   <header class="site-header">
     <div class="title-logo">Rx Reflection</div>
-    <div class="menu-icon"></div>
     {{#if logged_in}}
     <nav>
       <button class="btn" id="logout">logout</button>
-      {{#unless patientUser}}
-      <a href="/add-patient" class="nav-link">Add Patient</a>
-      {{/unless}}
     </nav>
     {{/if}}
   </header>
